refactored to include s3
- moved backend handling to separate modules
- added s3/rclone backend for mounting buckets
- updated README to show more info about ssh tunneling, s3, rclone, etc
- logging, break_handler moved to util module
- added diagram for tunneling but not sure where to store it yet
